+++ autode/config.py
@@ -1,11 +1,11 @@
 import os
-import autode.wrappers.implicit_solvent_types as solv
+from autode.wrappers import implicit_solvent_types as solv
 from autode.values import Frequency, Distance, Allocation
 from autode.wrappers.keywords import KeywordsSet, MaxOptCycles
-from autode.wrappers.basis_sets import def2svp, def2tzvp, def2ecp, def2tzecp
-from autode.wrappers.functionals import pbe0
-from autode.wrappers.dispersion import d3bj
-from autode.wrappers.ri import rijcosx
+from autode.wrappers.keywords.basis_sets import def2svp, def2tzvp, def2ecp, def2tzecp
+from autode.wrappers.keywords.functionals import pbe0
+from autode.wrappers.keywords.dispersion import d3bj
+from autode.wrappers.keywords.ri import rijcosx
 location = os.path.abspath(__file__)