+++ autode/wrappers/__init__.py
@@ -1,0 +1,6 @@
+from autode.wrappers.keywords import (basis_sets,
+                                      dispersion,
+                                      functionals,
+                                      implicit_solvent_types,
+                                      ri,
+                                      wf)