+++ autode/wrappers/keywords/dispersion.py
@@ -1,4 +1,4 @@
-from autode.wrappers.keywords import DispersionCorrection
+from autode.wrappers.keywords.keywords import DispersionCorrection
 
 d3bj = DispersionCorrection(name='d3bj',
                             doi_list=['10.1063/1.3382344', '10.1002/jcc.21759'],