+++ autode/wrappers/keywords/ri.py
@@ -1,4 +1,4 @@
-from autode.wrappers.keywords import RI
+from autode.wrappers.keywords.keywords import RI
 
 rijcosx = RI(name='rijcosx',
              DOIs=['10.1016/j.chemphys.2008.10.036'],