+++ doc/changelog.rst
@@ -24,6 +24,7 @@ Usability improvements/Changes
 * Removes :code:`autode.calculations.Calculation.print_final_output_lines` in favour of a method on :code:`calculation.output`
 * Makes many methods in :code:`autode.calculations.Calculation` private
 * Deprecates all :code:`autode.calculations.Calculation.get_<method>` methods in favour of properties
+* Adds an :code:`autode.wrappers.keywords` package to improve file structure
 
 
 Functionality improvements